+++ wx/lib/agw/aui/framemanager.py
@@ -3761,7 +3761,7 @@ def GetNotebookRoot(panes, notebook_id):
 
 def EscapeDelimiters(s):
     """
-    Changes ``;`` into ``\`` and ``|`` into ``\|`` in the input string.
+    Changes ``;`` into ``\\`` and ``|`` into ``|\\`` in the input string.
 
     :param string `s`: the string to be analyzed.