+++ CHANGES.rst
@@ -28,6 +28,13 @@ New and improved in this release:
   Python environment used for the build. A dependency has been added to
   requirements/devel.txt to help ensure that the correct version is installed.
 
+* Change wx.App.InitLocale to just do `locale.setlocale(locale.LC_ALL, "C")`
+  to undo what Python (3.8+ on Windows) does. This lets wxWidgets start with an
+  uninitialized locale as it expects. (#1637)
+
+* Fixed issues related to `time_t` being treated as a 32-bit value on Windows.
+  (#1910)
+
 
 
 4.1.1 "An attitude of gratitude"